Minor clean-ups
